[Refactor] Refactor alert module. (#3023)
* [Refactor] Refactor alert module.

* address comment

+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/entity/FlinkCluster.java
@@ -110,7 +110,7 @@ public class FlinkCluster implements Serializable {
   private Date endTime;
 
   @TableField(updateStrategy = FieldStrategy.IGNORED)
-  private Integer alertId;
+  private Long alertId;
 
   private transient Integer allJobs = 0;
